automatic_fraud_detection/training/requirements.txt
mlflow
scikit-learn
pandas
boto3
pytest
python-dotenv
psycopg2-binary